backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "katsuyoriiii"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi[all] (>=0.117.1,<0.118.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "fastapi-filter[sqlalchemy] (>=2.0.1,<3.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
